Cleanup of Dockerfiles and make_smi_build
Extend the set of OpenPegasjs build variables for the server image build
and document them.  Note that many are commented out so that their
default value applies

Remove most of the configuration environment variables from the server
dockerfile since they are not needed.

Add a new section to the makefile (provision server) which is used to
provision the repository and any test/other repositories.   Today it
simple executes the testrepository target in pegasus/makefile.